i forgor to commit yikes

+++ FA-backend/src/FitnessApp.InterfaceLayer/Dtos/Classes/Walk/CreateWalkDto.cs
@@ -5,9 +5,9 @@ public class CreateWalkDto
     public int _user_id { get; set; }
     public int _distance { get; set; }
 
-    public int _startLat { get; set; }
-    public int _startLong { get; set; }
+    public float _startLat { get; set; }
+    public float _startLong { get; set; }
     
-    public int _endLat { get; set; }
-    public int _endLong { get; set; }
+    public float _endLat { get; set; }
+    public float _endLong { get; set; }
 }
